+++ plugins/css-has-pseudo/src/index.ts
@@ -30,8 +30,8 @@ const creator: PluginCreator<{ preserve?: boolean, specificityMatchingName?: str
 				let selectorAST;
 				try {
 					selectorAST = parser().astSync(selector);
-				} catch (_) {
-					rule.warn(result, `Failed to parse selector : ${selector}`);
+				} catch (err) {
+					rule.warn(result, `Failed to parse selector : "${selector}" with message: "${err.message}"`);
 					return selector;
 				}