+++ plugins/postcss-attribute-case-insensitive/.tape.mjs
@@ -5,6 +5,10 @@ postcssTape(plugin)({
 	basic: {
 		message: "supports basic usage",
 	},
+	'invalid-selector': {
+		message: 'warns on invalid selectors',
+		warnings: 1
+	},
 	'examples/example': {
 		message: 'minimal example',
 	},