+++ plugins/postcss-nesting/src/lib/rule-within-rule.ts
@@ -1,15 +1,20 @@
 import shiftNodesBeforeParent from './shift-nodes-before-parent.js';
 import cleanupParent from './cleanup-parent.js';
 import mergeSelectors from './merge-selectors/merge-selectors.js';
-import type { Rule } from 'postcss';
+import type { Result, Rule } from 'postcss';
 import { options } from './options.js';
 
-export default function transformRuleWithinRule(node: Rule, parent: Rule, opts: options) {
+export default function transformRuleWithinRule(node: Rule, parent: Rule, result: Result, opts: options) {
 	// move previous siblings and the node to before the parent
 	shiftNodesBeforeParent(node, parent);
 
 	// update the selectors of the node to be merged with the parent
-	node.selectors = mergeSelectors(parent.selectors, node.selectors, opts);
+	try {
+		node.selectors = mergeSelectors(parent.selectors, node.selectors, opts);
+	} catch (err) {
+		node.warn(result, `Failed to parse selectors : "${parent.selector}" / "${node.selector}" with message: "${err.message}"`);
+		return;
+	}
 
 	// merge similar rules back together
 	const areSameRule = (node.type === 'rule' && parent.type === 'rule' && node.selector === parent.selector);